Minor refactoring Adding lombok to the project
