+++ pom.xml
@@ -15,6 +15,7 @@
     <junit.jupiter.api.version>5.10.1</junit.jupiter.api.version>
     <junit.jupiter.engine.version>5.10.1</junit.jupiter.engine.version>
     <assertj-core.version>3.25.1</assertj-core.version>
+    <lombok.version>1.18.30</lombok.version>
   </properties>
   <dependencies>
     <dependency>
@@ -34,6 +35,12 @@
       <artifactId>assertj-core</artifactId>
       <version>${assertj-core.version}</version>
       <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.projectlombok</groupId>
+      <artifactId>lombok</artifactId>
+      <version>${lombok.version}</version>
+      <scope>provided</scope>
     </dependency>
   </dependencies>
   <build>
