--- src/Gender.java
@@ -1,15 +1,0 @@
-public enum Gender {
-
-  MALE("Male"),
-  FEMALE("Female");
-
-  private final String value;
-
-  Gender(String value) {
-    this.value = value;
-  }
-
-  public String getValue() {
-    return value;
-  }
-}
