+++ src/main/java/model/Gender.java
@@ -1,0 +1,15 @@
+package model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public enum Gender {
+
+  MALE("Male"),
+  FEMALE("Female");
+
+  private final String value;
+
+}
